requirements.txt
Django==5.2.8
djangorestframework==3.21.0
django-filter==23.2
